MAINT: Make config_manager generic
